pyproject.toml
[project]
name = "odoo-qa"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "allure-pytest>=2.15.3",
    "playwright>=1.56.0",
    "pytest>=9.0.2",
    "pytest-html>=4.1.1",
    "pytest-xdist>=3.8.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
